Trait ndarray::blas::AsBlas

pub trait AsBlas<A, S, D> {
    fn blas_checked(&mut self) -> Result<BlasArrayViewMut<A, D>, ShapeError> where S: DataOwned + DataMut, A: Clone;

    fn blas(&mut self) -> BlasArrayViewMut<A, D> where S: DataOwned<Elem=A> + DataMut, A: Clone { ... }
}

Convert an array into a blas friendly wrapper.

Note that blas suppors four different element types: f32, f64, Complex<f32>, and Complex<f64>.

Requires features = "rblas"

Required Methods

fn blas_checked(&mut self) -> Result<BlasArrayViewMut<A, D>, ShapeError> where S: DataOwned + DataMut, A: Clone

Return an array view implementing Vector (1D) or Matrix (2D) traits.

Elements are copied if needed to produce a contiguous matrix.
The result is always mutable, due to the requirement of having write access to update the layout either way. Breaks sharing if the array is an Array.

Errors: Produces an error if any dimension is larger than c_int::MAX.

Provided Methods

fn blas(&mut self) -> BlasArrayViewMut<A, D> where S: DataOwned<Elem=A> + DataMut, A: Clone

Equivalent to .blas_checked().unwrap()

Panics if there was a an error in blas_checked.