Enum ndarray::CowRepr

pub enum CowRepr<'a, A> {
    View(ViewRepr<&'a A>),
    Owned(OwnedRepr<A>),
}

CowArray's representation.

Don't use this type directly—use the type alias CowArray for the array type!

Variants

View(ViewRepr<&'a A>)

Borrowed data.

Owned(OwnedRepr<A>)

Owned data.

Implementations

impl<'a, A> CowRepr<'a, A>

pub fn is_view(&self) -> bool

Returns true iff the data is the View variant.

pub fn is_owned(&self) -> bool

Returns true iff the data is the Owned variant.