Function ndarray::linalg::cholesky

pub fn cholesky<A: ComplexField>(a: Mat<A>) -> Mat<A>

Factor a = L LT.

a should be a square matrix, hermitian and positive definite.

https://en.wikipedia.org/wiki/Cholesky_decomposition

“The Cholesky decomposition is mainly used for the numerical solution of linear equations Ax = b.

If A is symmetric and positive definite, then we can solve Ax = b by first computing the Cholesky decomposition A = LL*, then solving Ly = b for y by forward substitution, and finally solving L*x = y for x by back substitution.”

Return L.