ndarray::linalg

Function general_mat_mul

pub fn general_mat_mul<A, S1, S2, S3>(
    alpha: A,
    a: &ArrayBase<S1, Ix2>,
    b: &ArrayBase<S2, Ix2>,
    beta: A,
    c: &mut ArrayBase<S3, Ix2>,
)
where
    S1: Data<Elem = A>,
    S2: Data<Elem = A>,
    S3: DataMut<Elem = A>,
    A: LinalgScalar,

General matrix-matrix multiplication.

Compute C ← α A B + β C

The array shapes must agree in the way that if a is M × N, then b is N × K and c is M × K.

Panics if array shapes are not compatible
Note: If enabled, uses blas gemm for elements of f32, f64 when memory layout allows. The default matrixmultiply backend is otherwise used for f32, f64 for all memory layouts.