Struct ndarray::Axis

pub struct Axis(pub usize);

An axis index.

An axis one of an array’s “dimensions”; an n-dimensional array has n axes. Axis 0 is the array’s outermost axis and n-1 is the innermost.

All array axis arguments use this type to make the code easier to write correctly and easier to understand.

Methods

impl Axis

fn index(&self) -> usize

Return the index of the axis.