ndarray

Struct RawRef

pub struct RawRef<A, D>(/* private fields */);

A reference to an n-dimensional array whose data is not safe to read or write.

This type is similar to ArrayRef but does not guarantee that its data is safe to read or write; i.e., the underlying data may come from a shared array or be otherwise unsafe to dereference. This type should be used sparingly and with extreme caution; most of its methods either provide pointers or return RawArrayView, both of which tend to be full of unsafety.

For the few times when this type is appropriate, it has the same AsRef semantics as LayoutRef; see its documentation on writing functions for information on how to properly handle functionality on this type.

Implementations

impl<A, D: Dimension> RawRef<A, D>

pub fn get_ptr<I>(&self, index: I) -> Option<*const A>
where I: NdIndex<D>,

Return a raw pointer to the element at index, or return None if the index is out of bounds.

use ndarray::arr2;

let a = arr2(&[[1., 2.], [3., 4.]]);

let v = a.raw_view();
let p = a.get_ptr((0, 1)).unwrap();

assert_eq!(unsafe { *p }, 2.);

impl<A, D: Dimension> RawRef<A, D>

pub fn get_mut_ptr<I>(&mut self, index: I) -> Option<*mut A>
where I: NdIndex<D>,

Return a raw pointer to the element at index, or return None if the index is out of bounds.

use ndarray::arr2;

let mut a = arr2(&[[1., 2.], [3., 4.]]);

let v = a.raw_view_mut();
let p = a.get_mut_ptr((0, 1)).unwrap();

unsafe {
    *p = 5.;
}

assert_eq!(a.get((0, 1)), Some(&5.));

impl<A, D: Dimension> RawRef<A, D>

pub fn as_ptr(&self) -> *const A

Return a pointer to the first element in the array.

Raw access to array elements needs to follow the strided indexing scheme: an element at multi-index I in an array with strides S is located at offset

Σ0 ≤ k < d Ik × Sk

where d is self.ndim().

pub fn as_mut_ptr(&mut self) -> *mut A

Return a mutable pointer to the first element in the array reference.

impl<A, D: Dimension> RawRef<A, D>

pub fn raw_view(&self) -> RawArrayView<A, D>

Return a raw view of the array.

pub fn raw_view_mut(&mut self) -> RawArrayViewMut<A, D>

Return a raw mutable view of the array.