Struct ndarray::Si

pub struct Si(pub Ixs, pub Option<Ixs>, pub Ixs);

A slice, a description of a range of an array axis.

Fields are begin, end and stride, where negative begin or end indexes are counted from the back of the axis.

If end is None, the slice extends to the end of the axis.

Examples

Si(0, None, 1) is the full range of an axis. Python equivalent is [:].

Si(a, Some(b), 2) is every second element from a until b. Python equivalent is [a:b:2].

Si(a, None, -1) is every element, in reverse order, from a until the end. Python equivalent is [a::-1]