Struct ndarray::SliceInfo

#[repr(C)]

pub struct SliceInfo<T: ?Sized, D: Dimension> { /* fields omitted */ }

Represents all of the necessary information to perform a slice.

The type T is typically [SliceOrIndex; n], [SliceOrIndex], or Vec<SliceOrIndex>. The type D is the output dimension after calling .slice().

Methods

impl<T, D> SliceInfo<T, D> where
T: AsRef<[SliceOrIndex]>,
D: Dimension,

pub fn new(indices: T) -> Result<SliceInfo<T, D>, ShapeError>

Returns a new SliceInfo instance.

Errors if D is not consistent with indices.

impl<T: ?Sized, D> SliceInfo<T, D> where
T: AsRef<[SliceOrIndex]>,
D: Dimension,

pub fn out_ndim(&self) -> usize

Returns the number of dimensions after calling .slice() (including taking subviews).

If D is a fixed-size dimension type, then this is equivalent to D::NDIM.unwrap(). Otherwise, the value is calculated by iterating over the ranges/indices.