Trait ndarray::RawData

pub unsafe trait RawData: Sized {
    type Elem;
    fn __private__(&self) -> PrivateMarker;
}

Array representation trait.

For an array that meets the invariants of the ArrayBase type. This trait does not imply any ownership or lifetime; pointers to elements in the array may not be safe to dereference.

Note: RawData is not an extension interface at this point. Traits in Rust can serve many different roles. This trait is public because it is used as a bound on public methods.

Associated Types

type Elem

The array element type.

Required methods

fn __private__(&self) -> PrivateMarker

This trait is private to implement; this method exists to make it impossible to implement outside the crate.