ndarray

Trait ShapeBuilder

pub trait ShapeBuilder {
    type Dim: Dimension;
    type Strides;

    // Required methods
    fn into_shape_with_order(self) -> Shape<Self::Dim>;
    fn f(self) -> Shape<Self::Dim>;
    fn set_f(self, is_f: bool) -> Shape<Self::Dim>;
    fn strides(self, strides: Self::Strides) -> StrideShape<Self::Dim>;
}

A trait for Shape and D where D: Dimension that allows customizing the memory layout (strides) of an array shape.

This trait is used together with array constructor methods like Array::from_shape_vec.

Required Associated Types

type Dim: Dimension

The type that captures the built shape’s dimensionality.

type Strides

The type that captures the built shape’s stride pattern.

Required Methods

fn into_shape_with_order(self) -> Shape<Self::Dim>

Turn into a contiguous-element Shape.

fn f(self) -> Shape<Self::Dim>

Convert into a Fortran-order (a.k.a., column-order) Shape.

fn set_f(self, is_f: bool) -> Shape<Self::Dim>

Set the order of the shape to either Fortran or non-Fortran.

fn strides(self, strides: Self::Strides) -> StrideShape<Self::Dim>

Set the strides of the shape.