Type Definition ndarray::RcArray

type RcArray<A, D> = ArrayBase<OwnedRcRepr<A>, D>;

An array where the data has shared ownership and is copy on write. It can act as both an owner as the data as well as a shared reference (view like).

The RcArray<A, D> is parameterized by A for the element type and D for the dimensionality.

ArrayBase is used to implement both the owned arrays and the views; see its docs for an overview of all array features.

See also:

Constructor Methods for Owned Arrays

Methods For All Array Types