Crate ndarray

The ndarray crate provides an n-dimensional container similar to numpy's ndarray.

ArrayBase
The n-dimensional array type itself, parameterized by data storage.

Array
Array where the data is reference counted and copy on write, it can act as both an owner as the data as well as a lightweight view.

OwnedArray
Array where the data is owned uniquely.

ArrayView
A lightweight array view.

ArrayViewMut
A lightweight read-write array view.

Crate Summary and Status

Implements the numpy striding and broadcasting scheme for n-dimensional arrays

Array is clone on write, so it can be both a view or an owner of the data. OwnedArray is a uniquely owned array.

Focus is on being a generic n-dimensional container

Due to iterators, arithmetic operations, matrix multiplication etc are not very well optimized, this is not a serious crate for numerics or linear algebra.

There is no integration with linear algebra packages (at least not yet).

Crate feature flags

assign_ops

Optional, requires nightly

Enables the compound assignment operators

rustc-serialize

Optional, stable

Enables serialization support

Modules

linalg
A few linear algebra operations on two-dimensional arrays.

Structs

ArrayBase
The Array type is an N-dimensional array.

Elements
An iterator over the elements of an array.

ElementsMut
An iterator over the elements of an array.

Indexed
An iterator over the indexes and elements of an array.

IndexedMut
An iterator over the indexes and elements of an array.

Indexes
An iterator of the indexes of an array shape.

Si
A slice, a description of a range of an array axis.

Constants

S
Slice value for the full range of an axis.

Traits

Data
Array's inner representation.

DataClone
Clone an Array's storage.

DataMut
Array's writable inner representation.

DataOwned
Array representation that is a unique or shared owner of its data.

DataShared
Array representation that is a lightweight view.

Dimension
Trait for the shape and index types of arrays.

FixedInitializer
Fixed-size array used for array initialization

Initializer
Slice or fixed-size array used for array initialization

RemoveAxis
Helper trait to define a larger-than relation for array shapes: removing one axis from Self gives smaller dimension Smaller.

SliceRange

Functions

arr0
Return a zero-dimensional array with the element x.

arr1
Return a one-dimensional array with elements from xs.

arr2
Return a two-dimensional array with elements from xs.

arr3
Return a three-dimensional array with elements from xs.

aview0
Return a zero-dimensional array view borrowing x.

aview1
Return a one-dimensional array view with elements borrowing xs.

aview2
Return a two-dimensional array view with elements borrowing xs.

Type Definitions

Array
Array where the data is reference counted and copy on write, it can act as both an owner as the data as well as a lightweight view.

ArrayView
A lightweight array view.

ArrayViewMut
A lightweight read-write array view.

Ix
Array index type

Ixs
Array index type (signed)

OwnedArray
Array where the data is owned uniquely.