Crate ndarray

The ndarray crate provides the Array type, an n-dimensional container similar to numpy's ndarray.

Modules

linalg
A few linear algebra operations on two-dimensional arrays.

Structs

Array
The Array type is an N-dimensional array.

Elements
An iterator over the elements of an array.

ElementsMut
An iterator over the elements of an array.

IndexedElements
An iterator over the indexes and elements of an array.

IndexedElementsMut
An iterator over the indexes and elements of an array.

Indexes
An iterator of the indexes of an array shape.

Si
A slice, a description of a range of an array axis.

Constants

S
Slice value for the full range of an axis.

Traits

ArrInit
Slice or fixed-size array used for array initialization

Dimension
Trait for the shape and index types of arrays.

RemoveAxis
Helper trait to define a larger-than relation for array shapes: removing one axis from Self gives smaller dimension Smaller.

Functions

arr0
Return a zero-dimensional array with the element x.

arr1
Return a one-dimensional array with elements from xs.

arr2
Return a two-dimensional array with elements from xs.

ixrange
Like range, except with array indexes.

Type Definitions

Ix
Array index type

Ixs
Array index type (signed)