Crate ndarray

The ndarray crate provides the Array type, an n-dimensional numerical container similar to numpy's ndarray.

Modules

linalg
A few linear algebra operations on two-dimensional arrays.

Macros

matrix!
Create a 2D matrix

Structs

Array
The Array type is an N-dimensional array.

Elements
An iterator over the elements of an array.

ElementsMut
An iterator over the elements of an array.

IndexedElements
An iterator over the indexes and elements of an array.

IndexedElementsMut
An iterator over the indexes and elements of an array.

Indexes
An iterator of the indexes of an array shape.

Si
A slice, a description of a range of an array axis.

Statics

S
Slice value for the full range of an axis.

Traits

Dimension
Trait for the shape and index types of arrays.

RemoveAxis
Helper trait to define a larger-than relation for array shapes: removing one axis from Self gives smaller dimension E.

Functions

arr0
Return a zero-dimensional array with the element x.

arr1
Return a one-dimensional array with elements from xs.

arr2
Return a two-dimensional array with elements from xs.

d1
Construct one-dimensional array shape. Helper function to use where integer literal inference isn't working.

d2
Construct two-dimensional array shape. Helper function to use where integer literal inference isn't working.

d3
Construct three-dimensional array shape. Helper function to use where integer literal inference isn't working.

d4
Construct four-dimensional array shape. Helper function to use where integer literal inference isn't working.

ixrange
Like range, except with array indexes.

Type Definitions

Ix
Array index type

Ixs
Array index type (signed)