ndarray

Enum SliceInfoElem

pub enum SliceInfoElem {
    Slice {
        start: isize,
        end: Option<isize>,
        step: isize,
    },
    Index(isize),
    NewAxis,
}

A slice (range with step), an index, or a new axis token.

See also the s![] macro for a convenient way to create a SliceInfo<[SliceInfoElem; n], Din, Dout>.

Examples

SliceInfoElem::Index(a) is the index a. It can also be created with SliceInfoElem::from(a). The Python equivalent is [a]. The macro equivalent is s![a].

SliceInfoElem::Slice { start: 0, end: None, step: 1 } is the full range of an axis. It can also be created with SliceInfoElem::from(..). The Python equivalent is [:]. The macro equivalent is s![..].

SliceInfoElem::Slice { start: a, end: Some(b), step: 2 } is every second element from a until b. It can also be created with SliceInfoElem::from(Slice::from(a..b).step_by(2)). The Python equivalent is [a:b:2]. The macro equivalent is s![a..b;2].

SliceInfoElem::Slice { start: a, end: None, step: -1 } is every element, from a until the end, in reverse order. It can also be created with SliceInfoElem::from(Slice::from(a..).step_by(-1)). The Python equivalent is [a::-1]. The macro equivalent is s![a..;-1].

SliceInfoElem::NewAxis is a new axis of length 1. It can also be created with SliceInfoElem::from(NewAxis). The Python equivalent is [np.newaxis]. The macro equivalent is s![NewAxis].

Variants

Slice

A range with step size. end is an exclusive index. Negative start or end indexes are counted from the back of the axis. If end is None, the slice extends to the end of the axis.

Fields

start: isize

start index; negative are counted from the back of the axis

end: Option<isize>

end index; negative are counted from the back of the axis; when not present the default is the full length of the axis.

step: isize

step size in elements; the default is 1, for every element.

Index(isize)

A single index.

NewAxis

A new axis of length 1.

Implementations

impl SliceInfoElem

pub fn is_slice(&self) -> bool

Returns true if self is a Slice value.

pub fn is_index(&self) -> bool

Returns true if self is an Index value.

pub fn is_new_axis(&self) -> bool

Returns true if self is a NewAxis value.