Module ndarray::iter

Producers, iterables and iterators.

This module collects all concrete producer, iterable and iterator implementation structs.

See also NdProducer.

Structs

Axes
An iterator over the length and stride of each axis of an array.

AxisChunksIter
An iterator that traverses over the specified axis and yields views of the specified size on this axis.

AxisChunksIterMut
An iterator that traverses over the specified axis and yields mutable views of the specified size on this axis.

AxisIter
An iterator that traverses over an axis and and yields each subview.

AxisIterMut
An iterator that traverses over an axis and and yields each subview (mutable)

ExactChunks
Exact chunks producer and iterable.

ExactChunksIter
Exact chunks iterator.

ExactChunksIterMut
Exact chunks iterator.

ExactChunksMut
Exact chunks producer and iterable.

IndexedIter
An iterator over the indexes and elements of an array.

IndexedIterMut
An iterator over the indexes and elements of an array (mutable).

Indices
Indices producer and iterable.

IndicesIter
An iterator over the indexes of an array shape.

Iter
An iterator over the elements of an array.

IterMut
An iterator over the elements of an array (mutable).

Lanes
See .lanes() for more information.

LanesIter
An iterator that traverses over all dimensions but the innermost, and yields each inner row.

LanesIterMut
An iterator that traverses over all dimensions but the innermost, and yields each inner row (mutable).

LanesMut
See .lanes_mut() for more information.

Windows
Window producer and iterable