ndarray

Module prelude

ndarray prelude.

This module contains the most used types, type aliases, traits, functions, and macros that you can import easily as a group.

use ndarray::prelude::*;

Re-exports

pub use crate::ArcArray;

pub use crate::Array;

pub use crate::ArrayBase;

pub use crate::ArrayRef;

pub use crate::ArrayView;

pub use crate::ArrayViewMut;

pub use crate::CowArray;

pub use crate::LayoutRef;

pub use crate::RawArrayView;

pub use crate::RawArrayViewMut;

pub use crate::RawRef;

pub use crate::Axis;

pub use crate::Dim;

pub use crate::Dim;

pub use crate::Dimension;

pub use crate::ArrayRef0;

pub use crate::ArrayRef1;

pub use crate::ArrayRef2;

pub use crate::ArrayRef3;

pub use crate::ArrayRef4;

pub use crate::ArrayRef5;

pub use crate::ArrayRef6;

pub use crate::ArrayRefD;

pub use crate::Array0;

pub use crate::Array1;

pub use crate::Array2;

pub use crate::Array3;

pub use crate::Array4;

pub use crate::Array5;

pub use crate::Array6;

pub use crate::ArrayD;

pub use crate::ArrayView0;

pub use crate::ArrayView1;

pub use crate::ArrayView2;

pub use crate::ArrayView3;

pub use crate::ArrayView4;

pub use crate::ArrayView5;

pub use crate::ArrayView6;

pub use crate::ArrayViewD;

pub use crate::ArrayViewMut0;

pub use crate::ArrayViewMut1;

pub use crate::ArrayViewMut2;

pub use crate::ArrayViewMut3;

pub use crate::ArrayViewMut4;

pub use crate::ArrayViewMut5;

pub use crate::ArrayViewMut6;

pub use crate::ArrayViewMutD;

pub use crate::Ix0;

pub use crate::Ix0;

pub use crate::Ix1;

pub use crate::Ix1;

pub use crate::Ix2;

pub use crate::Ix2;

pub use crate::Ix3;

pub use crate::Ix3;

pub use crate::Ix4;

pub use crate::Ix4;

pub use crate::Ix5;

pub use crate::Ix5;

pub use crate::Ix6;

pub use crate::Ix6;

pub use crate::IxDyn;

pub use crate::IxDyn;

pub use crate::arr0;

pub use crate::arr1;

pub use crate::arr2;

pub use crate::aview0;

pub use crate::aview1;

pub use crate::aview2;

pub use crate::aview_mut1;

pub use crate::ShapeBuilder;

pub use crate::NewAxis;

pub use crate::AsArray;

pub use crate::NdFloat;

Macros

array

Create an Array with one, two, three, four, five, or six dimensions.

azip

Array zip macro: lock step function application across several arrays and producers.

s

Slice argument constructor.