Module ndarray::prelude

ndarray prelude.

This module contains the most used types, type aliases, traits and functions that you can import easily as a group.

extern crate ndarray;

use ndarray::prelude::*;

Reexports

pub use ArrayBase;

pub use Array;

pub use RcArray;

pub use ArrayView;

pub use ArrayViewMut;

pub use Axis;

pub use Dim;

pub use Ix;

pub use Ixs;

pub use Dimension;

pub use Array0;

pub use Array1;

pub use Array2;

pub use Array3;

pub use Array4;

pub use Array5;

pub use Array6;

pub use ArrayD;

pub use ArrayView0;

pub use ArrayView1;

pub use ArrayView2;

pub use ArrayView3;

pub use ArrayView4;

pub use ArrayView5;

pub use ArrayView6;

pub use ArrayViewD;

pub use ArrayViewMut0;

pub use ArrayViewMut1;

pub use ArrayViewMut2;

pub use ArrayViewMut3;

pub use ArrayViewMut4;

pub use ArrayViewMut5;

pub use ArrayViewMut6;

pub use ArrayViewMutD;

pub use Ix0;

pub use Ix1;

pub use Ix2;

pub use Ix3;

pub use Ix4;

pub use Ix5;

pub use Ix6;

pub use IxDyn;

pub use arr0;

pub use arr1;

pub use arr2;

pub use aview0;

pub use aview1;

pub use aview2;

pub use aview_mut1;

pub use ShapeBuilder;

pub use NdFloat;

pub use AsArray;