Module ndarray::prelude

ndarray prelude.

This module contains the most used types, type aliases, traits and functions that you can import easily as a group.

extern crate ndarray;

use ndarray::prelude::*;

Reexports

pub use ::{ArrayBase, Array, RcArray, ArrayView, ArrayViewMut};

pub use OwnedArray;

pub use ::{Axis, Ix, Ixs, Dimension};

pub use ::{NdFloat, AsArray};

pub use ::{arr1, arr2, aview0, aview1, aview2};

pub use ::{ShapeBuilder};