Struct ndarray::Array

pub struct Array<A, D> {
    // some fields omitted
}

The Array type is an N-dimensional array.

A reference counted array with copy-on-write mutability.

The array is a container of numerical use, supporting all mathematical operators by applying them elementwise.

The array is both a view and a shared owner of its data. Some methods, for example slice(), merely change the view of the data, while methods like iadd() allow mutating the element values.

Calling a method for mutating elements, for example at_mut(), iadd() or iter_mut() will break sharing and require a clone of the data (if it is not uniquely held).

Method Conventions

Methods mutating the view or array elements in place use an i prefix, for example slice vs. islice and add vs iadd.

Broadcasting

Arrays support limited broadcasting, where arithmetic operations with array operands of different sizes can be carried out by repeating the elements of the smaller dimension array.

use ndarray::arr2;

let a = arr2::<f32>([[1., 2.],
                     [3., 4.]]);
let b = arr2::<f32>([[0., 1.]]);

let c = arr2::<f32>([[1., 3.],
                     [3., 5.]]);
// We can add because the shapes are compatible even if not equal.
assert!(
    c == a + b
);

Methods

impl<A: Clone + Zero, D: Dimension> Array<A, D>

fn zeros(dim: D) -> Array<A, D>

Construct an Array with zeros.

impl<A: Clone, D: Dimension> Array<A, D>

fn from_elem(dim: D, elem: A) -> Array<A, D>

Construct an Array with copies of elem.

impl<A> Array<A, Ix>

fn from_vec(v: Vec<A>) -> Array<A, Ix>

Create a one-dimensional array from a vector (no allocation needed).

fn from_iter<I: Iterator<A>>(it: I) -> Array<A, Ix>

Create a one-dimensional array from an iterator.

impl<A, D: Dimension> Array<A, D>

unsafe fn from_vec_dim(dim: D, v: Vec<A>) -> Array<A, D>

Create an array from a vector (with no allocation needed).

Unsafe because dimension is unchecked, and must be correct.

fn dim(&self) -> D

Return the shape of the array.

fn shape(&self) -> &[Ix]

Return the shape of the array as a slice.

fn is_standard_layout(&self) -> bool

Return true if the array data is laid out in contiguous “C order” where the last index is the most rapidly varying.

Return false otherwise, i.e the array is possibly not contiguous in memory, it has custom strides, etc.

fn raw_data<'a>(&'a self) -> &'a [A]

Return a slice of the array's backing data in memory order.

Note: Data memory order may not correspond to the index order of the array. Neither is the raw data slice is restricted to just the Array's view.

fn slice(&self, indexes: &[Si]) -> Array<A, D>

Return a sliced array.

Fail if indexes does not match the number of array axes.

fn islice(&mut self, indexes: &[Si])

Slice the array's view in place.

Fail if indexes does not match the number of array axes.

fn slice_iter<'a>(&'a self, indexes: &[Si]) -> Elements<'a, A, D>

Return an iterator over a sliced view.

Fail if indexes does not match the number of array axes.

fn at<'a>(&'a self, index: D) -> Option<&'a A>

Return a reference to the element at index, or return None if the index is out of bounds.

unsafe fn uchk_at<'a>(&'a self, index: D) -> &'a A

Perform unchecked array indexing.

Return a reference to the element at index.

Note: only unchecked for non-debug builds of ndarray.

unsafe fn uchk_at_mut<'a>(&'a mut self, index: D) -> &'a mut A

Perform unchecked array indexing.

Return a mutable reference to the element at index.

Note: Only unchecked for non-debug builds of ndarray.
Note: The array must be uniquely held when mutating it.

fn iter<'a>(&'a self) -> Elements<'a, A, D>

Return an iterator of references to the elements of the array.

Iterator element type is &'a A.

fn indexed_iter<'a>(&'a self) -> IndexedElements<'a, A, D>

Return an iterator of references to the elements of the array.

Iterator element type is (D, &'a A).

fn isubview(&mut self, axis: uint, index: uint)

Collapse dimension axis into length one, and select the subview of index along that axis.

Fail if index is past the length of the axis.

fn broadcast_iter<'a, E: Dimension>(&'a self, dim: E) -> Option<Elements<'a, A, E>>

Act like a larger size and/or shape array by broadcasting into a larger shape, if possible.

Return None if shapes can not be broadcast together.

Background

Two axes are compatible if they are equal, or one of them is 1.

In this instance, only the axes of the smaller side (self) can be 1.

Compare axes beginning with the last axis of each shape.

For example (1, 2, 4) can be broadcast into (7, 6, 2, 4) because its axes are either equal or 1 (or missing); while (2, 2) can not be broadcast into (2, 4).

The implementation creates an iterator with strides set to 0 for the axes that are to be repeated.

See broadcasting documentation for Numpy for more information.

fn swap_axes(&mut self, ax: uint, bx: uint)

Swap axes ax and bx.

Fail if the axes are out of bounds.

fn iter1d<'b>(&'b self, axis: uint, from: &D) -> Stride<'b, A>

One-dimensional iterator along axis, starting at from.

fn diag_iter<'a>(&'a self) -> Stride<'a, A>

Return an iterator over the diagonal elements of the array.

fn diag(&self) -> Array<A, Ix>

Return the diagonal as a one-dimensional array.

impl<A, D: RemoveAxis<E>, E: Dimension> Array<A, D>

fn subview(&self, axis: uint, index: uint) -> Array<A, E>

Select the subview index along axis and return an array with that axis removed.

Fail if index is past the length of the axis.

use ndarray::{arr1, arr2};

let a = arr2::<f32>([[1., 2.],
                     [3., 4.]]);

assert!(
    a.subview(0, 0) == arr1([1., 2.]) &&
    a.subview(1, 1) == arr1([2., 4.])
);

impl<A: Clone, D: Dimension> Array<A, D>

fn ensure_unique(&mut self)

Make the array unshared.

This method is mostly only useful with unsafe code.

fn at_mut<'a>(&'a mut self, index: D) -> Option<&'a mut A>

Return a mutable reference to the element at index, or return None if the index is out of bounds.

fn iter_mut<'a>(&'a mut self) -> ElementsMut<'a, A, D>

Return an iterator of mutable references to the elements of the array.

Iterator element type is &'a mut A.

fn indexed_iter_mut<'a>(&'a mut self) -> IndexedElementsMut<'a, A, D>

Return an iterator of indexes and mutable references to the elements of the array.

Iterator element type is (D, &'a mut A).

fn slice_iter_mut<'a>(&'a mut self, indexes: &[Si]) -> ElementsMut<'a, A, D>

Return an iterator of mutable references into the sliced view of the array.

Iterator element type is &'a mut A.

Fail if indexes does not match the number of array axes.

fn sub_iter_mut<'a>(&'a mut self, axis: uint, index: uint) -> ElementsMut<'a, A, D>

Select the subview index along axis and return an iterator of the subview.

Iterator element type is &'a mut A.

Fail if axis or index is out of bounds.

fn diag_iter_mut<'a>(&'a mut self) -> StrideMut<'a, A>

Return an iterator over the diagonal elements of the array.

fn raw_data_mut<'a>(&'a mut self) -> &'a mut [A]

Return a mutable slice of the array's backing data in memory order.

Note: Data memory order may not correspond to the index order of the array. Neither is the raw data slice is restricted to just the array's view.

Note: The data is uniquely held and nonaliased while it is mutably borrowed.

fn reshape<E: Dimension>(&self, shape: E) -> Array<A, E>

Transform the array into shape; any other shape with the same number of elements is accepted.

Fail if sizes are incompatible.

use ndarray::arr1;

arr1::<f32>([1., 2., 3., 4.]).reshape((2u, 2u));

fn assign<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise assigment to self from other.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn assign_scalar(&mut self, x: &A)

Perform an elementwise assigment to self from scalar x.

impl<A: Clone + Add<A, A>, D: RemoveAxis<E>, E: Dimension> Array<A, D>

fn sum(&self, axis: uint) -> Array<A, E>

Return sum along axis.

use ndarray::{arr0, arr1, arr2};

let a = arr2::<f32>([[1., 2.],
                     [3., 4.]]);
assert!(
    a.sum(0) == arr1([4., 6.]) &&
    a.sum(1) == arr1([3., 7.]) &&

    a.sum(0).sum(0) == arr0(10.)
);

Fail if axis is out of bounds.

impl<A: Clone + Field, D: RemoveAxis<E>, E: Dimension> Array<A, D>

fn mean(&self, axis: uint) -> Array<A, E>

Return mean along axis.

Fail if axis is out of bounds.

impl<A> Array<A, (Ix, Ix)>

fn row_iter<'a>(&'a self, index: uint) -> Stride<'a, A>

Return an iterator over the elements of row index.

Fail if index is out of bounds.

fn col_iter<'a>(&'a self, index: uint) -> Stride<'a, A>

Return an iterator over the elements of column index.

Fail if index is out of bounds.

impl<'a, A: Copy + Ring> Array<A, (Ix, Ix)>

fn mat_mul(&self, other: &Array<A, (Ix, Ix)>) -> Array<A, (Ix, Ix)>

Perform matrix multiplication of rectangular arrays self and other.

The array sizes must agree in the way that if self is M × N, then other is N × K.

Return a result array with shape M × K.

Fail if sizes are incompatible.

use ndarray::arr2;

let a = arr2::<f32>([[1., 2.],
                     [0., 1.]]);
let b = arr2::<f32>([[1., 2.],
                     [2., 3.]]);

assert!(
    a.mat_mul(&b) == arr2([[5., 8.],
                           [2., 3.]])
);

fn mat_mul_col(&self, other: &Array<A, Ix>) -> Array<A, Ix>

Perform the matrix multiplication of the rectangular array self and column vector other.

The array sizes must agree in the way that if self is M × N, then other is N.

Return a result array with shape M.

Fail if sizes are incompatible.

impl<A: Clone + Add<A, A>, D: Dimension> Array<A, D>

fn iadd<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise arithmetic operation between self and other, in place.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn iadd_scalar(&mut self, x: &A)

Perform an elementwise arithmetic operation between self and the scalar x, in place.

impl<A: Clone + Sub<A, A>, D: Dimension> Array<A, D>

fn isub<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise arithmetic operation between self and other, in place.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn isub_scalar(&mut self, x: &A)

Perform an elementwise arithmetic operation between self and the scalar x, in place.

impl<A: Clone + Mul<A, A>, D: Dimension> Array<A, D>

fn imul<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise arithmetic operation between self and other, in place.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn imul_scalar(&mut self, x: &A)

Perform an elementwise arithmetic operation between self and the scalar x, in place.

impl<A: Clone + Div<A, A>, D: Dimension> Array<A, D>

fn idiv<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise arithmetic operation between self and other, in place.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn idiv_scalar(&mut self, x: &A)

Perform an elementwise arithmetic operation between self and the scalar x, in place.

impl<A: Clone + Rem<A, A>, D: Dimension> Array<A, D>

fn irem<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise arithmetic operation between self and other, in place.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn irem_scalar(&mut self, x: &A)

Perform an elementwise arithmetic operation between self and the scalar x, in place.

impl<A: Clone + BitAnd<A, A>, D: Dimension> Array<A, D>

fn ibitand<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise arithmetic operation between self and other, in place.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn ibitand_scalar(&mut self, x: &A)

Perform an elementwise arithmetic operation between self and the scalar x, in place.

impl<A: Clone + BitOr<A, A>, D: Dimension> Array<A, D>

fn ibitor<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise arithmetic operation between self and other, in place.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn ibitor_scalar(&mut self, x: &A)

Perform an elementwise arithmetic operation between self and the scalar x, in place.

impl<A: Clone + BitXor<A, A>, D: Dimension> Array<A, D>

fn ibitxor<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise arithmetic operation between self and other, in place.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn ibitxor_scalar(&mut self, x: &A)

Perform an elementwise arithmetic operation between self and the scalar x, in place.

impl<A: Clone + Shl<A, A>, D: Dimension> Array<A, D>

fn ishl<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise arithmetic operation between self and other, in place.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn ishl_scalar(&mut self, x: &A)

Perform an elementwise arithmetic operation between self and the scalar x, in place.

impl<A: Clone + Shr<A, A>, D: Dimension> Array<A, D>

fn ishr<E: Dimension>(&mut self, other: &Array<A, E>)

Perform an elementwise arithmetic operation between self and other, in place.

If their shapes disagree, other is broadcast to the shape of self.

Fail if broadcasting isn't possible.

fn ishr_scalar(&mut self, x: &A)

Perform an elementwise arithmetic operation between self and the scalar x, in place.

impl<A: Clone + Neg<A>, D: Dimension> Array<A, D>

fn ineg(&mut self)

Perform an elementwise negation of self, in place.

impl<A: Clone + Not<A>, D: Dimension> Array<A, D>

fn inot(&mut self)

Perform an elementwise unary not of self, in place.