Struct ndarray::Si

pub struct Si(pub Ixs, pub Option<Ixs>, pub Ixs);

A slice, a description of a range of an array axis.

Fields are begin, end and stride, where negative begin or end indexes are counted from the back of the axis.

If end is None, the slice extends to the end of the axis.

See also the s![] macro, a convenient way to specify an array of Si.

Examples

Si(0, None, 1) is the full range of an axis. Python equivalent is [:]. Macro equivalent is s![..].

Si(a, Some(b), 2) is every second element from a until b. Python equivalent is [a:b:2]. Macro equivalent is s![a..b;2].

Si(a, None, -1) is every element, from a until the end, in reverse order. Python equivalent is [a::-1]. Macro equivalent is s![a..;-1].

The constant S is a shorthand for the full range of an axis.

Methods

impl Si

fn step(self, step: Ixs) -> Self