Trait ndarray::Data

pub unsafe trait Data: Sized {
    type Elem;
    fn __private__(&self) -> PrivateMarker;
}

Array representation trait.

Note: Data is not an extension interface at this point. Traits in Rust can serve many different roles. This trait is public because it is used as a bound on public methods.

Associated Types

type Elem

The array element type.

Required Methods

fn __private__(&self) -> PrivateMarker

This trait is private to implement; this method exists to make it impossible to implement outside the crate.