Trait ndarray::Dimension

pub trait Dimension: Clone + Eq + Debug + Send + Sync + Default + IndexMut<usize, Output = usize> + Add<Self, Output = Self> + AddAssign + for<'x> AddAssign<&'x Self> + Sub<Self, Output = Self> + SubAssign + for<'x> SubAssign<&'x Self> + Mul<usize, Output = Self> + Mul<Self, Output = Self> + MulAssign + for<'x> MulAssign<&'x Self> + MulAssign<usize> {
    type SliceArg: ?Sized + AsRef<[SliceOrIndex]>;
    type Pattern: IntoDimension<Dim = Self> + Clone + Debug + PartialEq + Eq + Default;
    type Smaller: Dimension;
    type Larger: Dimension + RemoveAxis;

    const NDIM: Option<usize>;

    fn ndim(&self) -> usize;

    fn into_pattern(self) -> Self::Pattern;

    fn zeros(ndim: usize) -> Self;

    fn __private__(&self) -> PrivateMarker;

    fn size(&self) -> usize { ... }

    fn size_checked(&self) -> Option<usize> { ... }

    fn as_array_view(&self) -> ArrayView1<'_, Ix> { ... }

    fn as_array_view_mut(&mut self) -> ArrayViewMut1<'_, Ix> { ... }

    fn into_dyn(self) -> IxDyn { ... }
}

Array shape and index trait.

This trait defines a number of methods and operations that can be used on dimensions and indices.

Note: This trait can not be implemented outside the crate

Associated Types

type SliceArg: ?Sized + AsRef<[SliceOrIndex]>

SliceArg is the type which is used to specify slicing for this dimension.

For the fixed size dimensions it is a fixed size array of the correct size, which you pass by reference. For the dynamic dimension it is a slice.

For Ix1: [SliceOrIndex; 1]

For Ix2: [SliceOrIndex; 2]

and so on..

For IxDyn: [SliceOrIndex]

The easiest way to create a &SliceInfo<SliceArg, Do> is using the s![] macro.

type Pattern: IntoDimension<Dim = Self> + Clone + Debug + PartialEq + Eq + Default

Pattern matching friendly form of the dimension value.

For Ix1: usize,

For Ix2: (usize, usize)

and so on..

For IxDyn: IxDyn

type Smaller: Dimension

Next smaller dimension (if applicable)

type Larger: Dimension + RemoveAxis

Next larger dimension

Loading content...
Associated Constants

const NDIM: Option<usize>

For fixed-size dimension representations (e.g. Ix2), this should be Some(ndim), and for variable-size dimension representations (e.g. IxDyn), this should be None.

Loading content...
Required methods

fn ndim(&self) -> usize

Returns the number of dimensions (number of axes).

fn into_pattern(self) -> Self::Pattern

Convert the dimension into a pattern matching friendly value.

fn zeros(ndim: usize) -> Self

Creates a dimension of all zeros with the specified ndim.

This method is useful for generalizing over fixed-size and variable-size dimension representations.

Panics if Self has a fixed size that is not ndim.

fn __private__(&self) -> PrivateMarker

This trait is private to implement; this method exists to make it impossible to implement outside the crate.

Loading content...
Provided methods

fn size(&self) -> usize

Compute the size of the dimension (number of elements)

fn size_checked(&self) -> Option<usize>

Compute the size while checking for overflow.

fn as_array_view(&self) -> ArrayView1<'_, Ix>

Borrow as a read-only array view.

fn as_array_view_mut(&mut self) -> ArrayViewMut1<'_, Ix>

Borrow as a read-write array view.

fn into_dyn(self) -> IxDyn

Convert the dimensional into a dynamic dimensional (IxDyn).

Loading content...