Trait ndarray::Dimension

pub trait Dimension: Clone + Eq {
    fn ndim(&self) -> uint;

    fn slice<'a>(&'a self) -> &'a [Ix] { ... }
    fn slice_mut<'a>(&'a mut self) -> &'a mut [Ix] { ... }
    fn size(&self) -> uint { ... }
    fn default_strides(&self) -> Self { ... }
    fn first_index(&self) -> Option<Self> { ... }
    fn next_for(&self, index: Self) -> Option<Self> { ... }
    fn stride_offset(index: &Self, strides: &Self) -> int { ... }
    fn stride_offset_checked(&self, strides: &Self, index: &Self) -> Option<int> { ... }
    fn do_slices(dim: &mut Self, strides: &mut Self, slices: &[Si]) -> int { ... }
}

Trait for the shape and index types of arrays.

Required Methods

fn ndim(&self) -> uint

Provided Methods

fn slice<'a>(&'a self) -> &'a [Ix]

fn slice_mut<'a>(&'a mut self) -> &'a mut [Ix]

fn size(&self) -> uint

fn default_strides(&self) -> Self

fn first_index(&self) -> Option<Self>

fn next_for(&self, index: Self) -> Option<Self>

Iteration -- Use self as size, and return next index after index or None if there are no more.

fn stride_offset(index: &Self, strides: &Self) -> int

Return stride offset for index.

fn stride_offset_checked(&self, strides: &Self, index: &Self) -> Option<int>

Return stride offset for this dimension and index.

fn do_slices(dim: &mut Self, strides: &mut Self, slices: &[Si]) -> int

Modify dimension, strides and return data pointer offset

Fail if slices does not correspond to the number of axes, if any stride is 0, or if any index is out of bounds.