Trait ndarray::NdFloat

pub trait NdFloat: Float + AddAssign + SubAssign + MulAssign + DivAssign + RemAssign + Display + Debug + LowerExp + UpperExp + ScalarOperand + LinalgScalar + Send + Sync { }

Floating-point element types f32 and f64.

Trait NdFloat is only implemented for f32 and f64 but encompasses as much float-relevant ndarray functionality as possible, including the traits needed for linear algebra and for right hand side scalar operations (ScalarOperand).

This trait can only be implemented by f32 and f64.