Trait ndarray::NdProducer

pub trait NdProducer {
    type Item;
    type Dim: Dimension;
    fn __private__(&self) -> PrivateMarker;
}

A producer of an n-dimensional set of elements; for example an array view, mutable array view or an iterator that yields chunks.

Producers are used as a arguments to Zip and azip!().

Associated Types

type Item

The element produced per iteration.

type Dim: Dimension

Dimension type

Required Methods

fn __private__(&self) -> PrivateMarker

This trait is private to implement; this method exists to make it impossible to implement outside the crate.