ndarray

Trait ShapeArg

pub trait ShapeArg {
    type Dim: Dimension;

    // Required method
    fn into_shape_and_order(self) -> (Self::Dim, Option<Order>);
}

Array shape argument with optional order parameter

Shape or array dimension argument, with optional Order parameter.

This is an argument conversion trait that is used to accept an array shape and (optionally) an ordering argument.

See for example .to_shape().

Required Associated Types

type Dim: Dimension

The type that captures the shape’s dimensionality.

Required Methods

fn into_shape_and_order(self) -> (Self::Dim, Option<Order>)

Convert the argument into a shape and an Order.