Trait ndarray::SliceArg

pub unsafe trait SliceArg<D: Dimension>: AsRef<[SliceInfoElem]> {
    type OutDim: Dimension;
    fn in_ndim(&self) -> usize;

    fn out_ndim(&self) -> usize;

    fn __private__(&self) -> PrivateMarker;
}

A type that can slice an array of dimension D.

This trait is unsafe to implement because the implementation must ensure that D, Self::OutDim, self.in_dim(), and self.out_ndim() are consistent with the &[SliceInfoElem] returned by self.as_ref() and that self.as_ref() always returns the same value when called multiple times.

Associated Types

type OutDim: Dimension

Dimensionality of the output array.

Loading content...
Required methods

fn in_ndim(&self) -> usize

Returns the number of axes in the input array.

fn out_ndim(&self) -> usize

Returns the number of axes in the output array.

fn __private__(&self) -> PrivateMarker

This trait is private to implement; this method exists to make it impossible to implement outside the crate.

Loading content...