ndarray

Type Alias ArcArray

pub type ArcArray<A, D> = ArrayBase<OwnedArcRepr<A>, D>;

An array where the data has shared ownership and is copy on write.

The ArcArray<A, D> is parameterized by A for the element type and D for the dimensionality.

It can act as both an owner as the data as well as a shared reference (view like). Calling a method for mutating elements on ArcArray, for example view_mut() or get_mut(), will break sharing and require a clone of the data (if it is not uniquely held).

ArcArray uses atomic reference counting like Arc, so it is Send and Sync (when allowed by the element type of the array too).

ArrayBase is used to implement both the owned arrays and the views; see its docs for an overview of all array features.

See also:

Constructor Methods for Owned Arrays

Methods For All Array Types

Aliased Type

pub struct ArcArray<A, D> { /* private fields */ }

Implementations

impl<A, D> ArcArray<A, D>
where D: Dimension,

Methods specific to ArcArray.

See also all methods for ArrayBase

pub fn is_unique(&self) -> bool

Returns true iff the inner Arc is not shared. If you want to ensure the Arc is not concurrently cloned, you need to provide a &mut self to this function.