Type Definition ndarray::Array

type Array<A, D> = ArrayBase<OwnedRepr<A>, D>;

An array that owns its data uniquely.

Array is the main n-dimensional array type, and it owns all its array elements.

The Array<A, D> is parameterized by A for the element type and D for the dimensionality.

ArrayBase is used to implement both the owned arrays and the views; see its docs for an overview of all array features.

See also:

Constructor Methods for Owned Arrays

Methods For All Array Types

Dimensionality-specific type alises Array1, Array2, Array3, ..., ArrayD, and so on.