Type Definition ndarray::ArrayView

type ArrayView<'a, A, D> = ArrayBase<ViewRepr<&'a A>, D>;

A read-only array view.

An array view represents an array or a part of it, created from an iterator, subview or slice of an array.

The ArrayView<'a, A, D> is parameterized by 'a for the scope of the borrow, A for the element type and D for the dimensionality.

Array views have all the methods of an array (see ArrayBase).

See also ArrayViewMut.

Methods

impl<'a, A, D> ArrayView<'a, A, D> where
D: Dimension,

Methods for read-only array views.

pub fn from_shape<Sh>(shape: Sh, xs: &'a [A]) -> Result<Self, ShapeError> where
Sh: Into<StrideShape<D>>,

Create a read-only array view borrowing its data from a slice.

Checks whether shape are compatible with the slice's length, returning an Err if not compatible.

use ndarray::ArrayView;
use ndarray::arr3;
use ndarray::ShapeBuilder;

let s = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12];
let a = ArrayView::from_shape((2, 3, 2).strides((1, 4, 2)),
                              &s).unwrap();

assert!(
    a == arr3(&[[[0, 2],
                 [4, 6],
                 [8, 10]],
                [[1, 3],
                 [5, 7],
                 [9, 11]]])
);
assert!(a.strides() == &[1, 4, 2]);

pub unsafe fn from_shape_ptr<Sh>(shape: Sh, ptr: *const A) -> Self where
Sh: Into<StrideShape<D>>,

Create an ArrayView<A, D> from shape information and a raw pointer to the elements.

Unsafe because caller is responsible for ensuring all of the following:

The elements seen by moving ptr according to the shape and strides must live at least as long as 'a and must not be not mutably aliased for the duration of 'a.

ptr must be non-null and aligned, and it must be safe to .offset() ptr by zero.

It must be safe to .offset() the pointer repeatedly along all axes and calculate the counts for the .offset() calls without overflow, even if the array is empty or the elements are zero-sized.

In other words,

All possible pointers generated by moving along all axes must be in bounds or one byte past the end of a single allocation with element type A. The only exceptions are if the array is empty or the element type is zero-sized. In these cases, ptr may be dangling, but it must still be safe to .offset() the pointer along the axes.

The offset in units of bytes between the least address and greatest address by moving along all axes must not exceed isize::MAX. This constraint prevents the computed offset, in bytes, from overflowing isize regardless of the starting point due to past offsets.

The offset in units of A between the least address and greatest address by moving along all axes must not exceed isize::MAX. This constraint prevents overflow when calculating the count parameter to .offset() regardless of the starting point due to past offsets.

The product of non-zero axis lengths must not exceed isize::MAX.

pub fn reborrow<'b>(self) -> ArrayView<'b, A, D> where
'a: 'b,

Convert the view into an ArrayView<'b, A, D> where 'b is a lifetime outlived by 'a'.

pub fn split_at(self, axis: Axis, index: Ix) -> (Self, Self)

Split the array view along axis and return one view strictly before the split and one view after the split.

Panics if axis or index is out of bounds.

Examples:

let a = array![[0, 1, 2, 3],
               [4, 5, 6, 7],
               [8, 9, 0, 1]];

The array a has two axes and shape 3 × 4:

         ─> Axis(1)
        ┌───┬───┬───┬───┐  0
      │ │ 0 │ 1 │ 2 │ 3 │
      v ├───┼───┼───┼───┤  1
 Axis(0)│ 4 │ 5 │ 6 │ 7 │
        ├───┼───┼───┼───┤  2
        │ 8 │ 9 │ 0 │ 1 │
        └───┴───┴───┴───┘  3 ↑
        0   1   2   3   4  ← possible split_at indices.

Row indices increase along Axis(0), and column indices increase along Axis(1). Note that we split “before” an element index, and that both 0 and the endpoint are valid split indices.

Example 1: Split a along the first axis, in this case the rows, at index 1.
This produces views v1 and v2 of shapes 1 × 4 and 2 × 4:

let (v1, v2) = a.split_at(Axis(0), 1);

        ┌───┬───┬───┬───┐       0  ↓ indices
      │ │ 0 │ 1 │ 2 │ 3 │ v1
      v └───┴───┴───┴───┘       1
 Axis(0)┌───┬───┬───┬───┐        
        │ 4 │ 5 │ 6 │ 7 │
        ├───┼───┼───┼───┤ v2    2
        │ 8 │ 9 │ 0 │ 1 │
        └───┴───┴───┴───┘       3

Example 2: Split a along the second axis, in this case the columns, at index 2.
This produces views u1 and u2 of shapes 3 × 2 and 3 × 2:

let (u1, u2) = a.split_at(Axis(1), 2);

            u1        u2
        0   1    2    3   4  indices →
        ┌───┬───┐ ┌───┬───┐
        │ 0 │ 1 │ │ 2 │ 3 │
        ├───┼───┤ ├───┼───┤
        │ 4 │ 5 │ │ 6 │ 7 │
        ├───┼───┤ ├───┼───┤
        │ 8 │ 9 │ │ 0 │ 1 │
        └───┴───┘ └───┴───┘
         ─> Axis(1)

pub fn into_slice(&self) -> Option<&'a [A]>

Return the array’s data as a slice, if it is contiguous and in standard order. Return None otherwise.